+++ R/utils_input.R
@@ -1,21 +1,23 @@
 #' Prepares Predictions
 #' 
-#' Converts predictions to vector, matrix or factor.
+#' Converts predictions to vector or matrix.
 #' 
 #' @noRd
 #' @keywords internal
 #' 
 #' @param x Object representing model predictions.
-#' @param ohe If `x` is a factor: should it be one-hot encoded? Default is `FALSE`.
-#' @returns Like `x`, but converted to matrix, vector, or factor.
-prepare_pred <- function(x, ohe = FALSE) {
+#' @returns Like `x`, but converted to matrix or vector.
+prepare_pred <- function(x) {
   if (is.data.frame(x) && ncol(x) == 1L) {
     x <- x[[1L]]
   }
-  if (ohe && is.factor(x)) {
-    return(fdummy(x))
+  if (!is.vector(x) || !is.matrix(x)) {
+    x <- as.matrix(x)
   }
-  if (is.vector(x) || is.matrix(x) || is.factor(x)) x else as.matrix(x)
+  if (!is.numeric(x)) {
+    stop("Predictions must be numeric!")
+  }
+  return(x)
 }
 
 #' Prepares Group BY Variable
@@ -99,7 +101,7 @@ prepare_w <- function(w, X) {
 #' 
 #' @returns A list with "y" (vector, matrix, or factor) and "y_names" (if `y`
 #'   was passed as column names).
-prepare_y <- function(y, X, ohe = FALSE) {
+prepare_y <- function(y, X) {
   if (NROW(y) < nrow(X) && all(y %in% colnames(X))) {
     y_names <- y
     if (is.data.frame(X) && length(y) == 1L) {
@@ -111,6 +113,6 @@ prepare_y <- function(y, X, ohe = FALSE) {
     stopifnot(NROW(y) == nrow(X))
     y_names <- NULL
   }
-  list(y = prepare_pred(y, ohe = ohe), y_names = y_names)
+  list(y = prepare_pred(y), y_names = y_names)
 }