+++ NEWS.md
@@ -1,3 +1,10 @@
+# hstats 1.2.0
+
+# Major changes
+
+- Factor-valued predictions are no longer possible.
+- Consequently, also removed "classification_error" loss.
+
 # hstats 1.1.2
 
 ## ICE plots